requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-tenants==3.6.1
psycopg2-binary
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.6.0
requests==2.32.3
redis==3.5.3
